+++ Packages/com.unity.cloud.gltfast/Runtime/Scripts/AssemblyInfo.cs
@@ -4,6 +4,7 @@
 using System.Runtime.CompilerServices;
 
 [assembly: InternalsVisibleTo("glTF-test-framework.Tests")] // TODO: Remove once those tests are no longer used.\
+[assembly: InternalsVisibleTo("glTFast.dots")]
 [assembly: InternalsVisibleTo("glTFast.Editor.Tests")]
 [assembly: InternalsVisibleTo("glTFast.Editor")]
 [assembly: InternalsVisibleTo("glTFast.Export")]